cleanup current ethereum plugin
